Add files via upload
I have to look up how to do this properly, but these three are the only files that I have created/changed so far. The product page is just a template so far and I added a route to access it via /product_page in App.js

the productPage.js is in the Components folder, productPage.js is in the styles folder.